pyproject.toml
[tool.poetry]
name = "experience_recorder"
version = "0.2.0"
description = ""
authors = ["frapercan <frapercan1@alum.us.es>"]

[tool.poetry.dependencies]
python = "^3.10"
docopt = "^0.6.2"
pyautogui = "^0.9.53"
pynput = "^1.7.6"
pyyaml = "^6.0"
pillow = "^9.4.0"
paddleocr = "^2.6.1.3"
paddlepaddle = "^2.4.2"
opencv-python = "4.5.5.64"
numpy = "^1.24.2"
sounddevice = "^0.4.6"
webdataset = "^0.2.48"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
coverage = {extras = ["toml"], version = "^7.2.1"}
flake8 = "^6.0.0"
flake8-bugbear = "^23.2.13"
taskipy = "^1.10.3"
sphinx = "^6.1.3"
sphinx-rtd-theme = "^1.2.0"
pytest-catchlog = "^1.2.2"





[tool.coverage.run]
omit = ["*/site-packages/*","experience_recorder/main.py"]

[tool.coverage.report]
fail_under = 100

[tool.taskipy.tasks]
html_docs = "make html -C docs"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

